:recycle: api: move ComplementaryCertification models to their domain
