+++ api/tests/certification/evaluation/unit/domain/models/PartnerCertificationScoring_test.js
@@ -1,6 +1,6 @@
-import { ObjectValidationError } from '../../../../src/shared/domain/errors.js';
-import { PartnerCertificationScoring } from '../../../../src/shared/domain/models/PartnerCertificationScoring.js';
-import { expect } from '../../../test-helper.js';
+import { ObjectValidationError } from '../../../../../../src/shared/domain/errors.js';
+import { PartnerCertificationScoring } from '../../../../../../src/shared/domain/models/index.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | PartnerCertificationScoring', function () {
   describe('constructor', function () {
